feat: implementing firestore batch!
