+++ src/BaseFirestoreBatchRepository.ts
@@ -1,36 +1,39 @@
-import { CollectionReference, WriteBatch } from '@google-cloud/firestore';
+import { WriteBatch, CollectionReference } from '@google-cloud/firestore';
 import { IEntity, WithOptionalId, Instantiable } from './types';
 import { getMetadataStorage, CollectionMetadata } from './MetadataStorage';
 import { serializeEntity } from './utils';
 
-// TODO: Eventually I'll abstract this class since right
-// now is tied to WriteBatch class from firestore
-export class FirestoreBatchRepository<T extends IEntity> {
-  private batch: WriteBatch;
-  private serializer: (item: T) => Object;
-  private colMetadata: CollectionMetadata;
-  private subColMetadata: CollectionMetadata[];
+export class BaseFirestoreBatchRepository<T extends IEntity> {
+  protected serializer: (item: T) => Object;
+  protected colMetadata: CollectionMetadata;
+  protected subColMetadata: CollectionMetadata[];
+  protected collectionPath: string;
+  protected colRef: CollectionReference;
 
   constructor(
-    private collection: CollectionReference,
-    private entity: Instantiable<T>
+    protected batch: WriteBatch,
+    protected entity: Instantiable<T>,
+    collectionPath?: string
   ) {
-    this.batch = collection.firestore.batch();
-
     const {
       getCollection,
       getSubCollection,
       getSubCollectionsFromParent,
+      firestoreRef,
     } = getMetadataStorage();
 
     this.colMetadata = getSubCollection(entity) || getCollection(entity);
     this.subColMetadata = getSubCollectionsFromParent(this.colMetadata.entity);
+
+    this.collectionPath = collectionPath || this.colMetadata.name;
+    this.colRef = firestoreRef.collection(this.collectionPath);
+
     this.serializer = (item: T) =>
       serializeEntity<T>(item, this.subColMetadata);
   }
 
   create = (item: WithOptionalId<T>) => {
-    const doc = item.id ? this.collection.doc(item.id) : this.collection.doc();
+    const doc = item.id ? this.colRef.doc(item.id) : this.colRef.doc();
     if (!item.id) {
       item.id = doc.id;
     }
@@ -39,14 +42,10 @@ export class FirestoreBatchRepository<T extends IEntity> {
   };
 
   update = (item: T) => {
-    this.batch.update(this.collection.doc(item.id), this.serializer(item));
+    this.batch.update(this.colRef.doc(item.id), this.serializer(item));
   };
 
   delete = (item: T) => {
-    this.batch.delete(this.collection.doc(item.id), this.serializer(item));
-  };
-
-  commit = () => {
-    return this.batch.commit();
+    this.batch.delete(this.colRef.doc(item.id), this.serializer(item));
   };
 }
