+++ test/functional/6-document-references.spec.ts
